Removing duplicate code in ArbitraryTransaction
