+++ Qora/src/qora/transaction/ArbitraryTransactionV1.java
@@ -4,32 +4,16 @@ import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 
-import ntp.NTP;
-
-import org.apache.commons.lang3.StringUtils;
-import org.json.simple.JSONObject;
-import org.json.simple.JSONValue;
-
-import qora.account.Account;
-import qora.account.PrivateKeyAccount;
-import qora.account.PublicKeyAccount;
-import qora.crypto.Base58;
-import qora.crypto.Crypto;
-import qora.naming.Name;
-import qora.payment.Payment;
-import qora.web.blog.BlogEntry;
-import utils.BlogUtils;
-import utils.StorageUtils;
-import api.BlogPostResource;
-
-import com.google.common.base.Charsets;
 import com.google.common.primitives.Bytes;
 import com.google.common.primitives.Ints;
 import com.google.common.primitives.Longs;
 
 import database.DBSet;
+import ntp.NTP;
+import qora.account.PrivateKeyAccount;
+import qora.account.PublicKeyAccount;
+import qora.crypto.Crypto;
 
 public class ArbitraryTransactionV1 extends ArbitraryTransaction {
 	protected static final int CREATOR_LENGTH = 32;
@@ -42,10 +26,6 @@ public class ArbitraryTransactionV1 extends ArbitraryTransaction {
 			+ REFERENCE_LENGTH + CREATOR_LENGTH + SERVICE_LENGTH
 			+ DATA_SIZE_LENGTH + FEE_LENGTH + SIGNATURE_LENGTH;
 
-	private PublicKeyAccount creator;
-	private int service;
-	private byte[] data;
-
 	public ArbitraryTransactionV1(PublicKeyAccount creator, int service,
 			byte[] data, BigDecimal fee, long timestamp, byte[] reference,
 			byte[] signature) {
@@ -54,18 +34,7 @@ public class ArbitraryTransactionV1 extends ArbitraryTransaction {
 		this.service = service;
 		this.data = data;
 		this.creator = creator;
-	}
-
-	// GETTERS/SETTERS
-
-	@Override
-	public int getService() {
-		return this.service;
-	}
-
-	@Override
-	public byte[] getData() {
-		return this.data;
+		this.payments = new ArrayList<>();
 	}
 
 	// PARSE CONVERT
@@ -124,20 +93,6 @@ public class ArbitraryTransactionV1 extends ArbitraryTransaction {
 
 		return new ArbitraryTransactionV1(creator, service, arbitraryData, fee,
 				timestamp, reference, signatureBytes);
-	}
-
-	@SuppressWarnings("unchecked")
-	@Override
-	public JSONObject toJson() {
-		// GET BASE
-		JSONObject transaction = this.getJsonBase();
-
-		// ADD CREATOR/SERVICE/DATA
-		transaction.put("creator", this.creator.getAddress());
-		transaction.put("service", this.service);
-		transaction.put("data", Base58.encode(this.data));
-
-		return transaction;
 	}
 
 	@Override
@@ -263,91 +218,6 @@ public class ArbitraryTransactionV1 extends ArbitraryTransaction {
 		return VALIDATE_OK;
 	}
 
-	// PROCESS/ORPHAN
-	@Override
-	public void process(DBSet db) {
-
-		try {
-			// NAME STORAGE UPDATE
-			if (service == 10) {
-				StorageUtils.processUpdate(getData(), signature, creator,
-						DBSet.getInstance());
-				StorageUtils.processUpdate(getData(), signature, creator, db);
-				// BLOGPOST?
-			} else if (service == 777) {
-				addToBlogMapOnDemand(DBSet.getInstance());
-				addToBlogMapOnDemand(db);
-			} else if (service == BlogUtils.COMMENT_SERVICE_ID) {
-				addToCommentMapOnDemand(DBSet.getInstance());
-				addToCommentMapOnDemand(db);
-			}
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-
-		// UPDATE CREATOR
-		this.creator.setConfirmedBalance(this.creator.getConfirmedBalance(db)
-				.subtract(this.fee), db);
-
-		// UPDATE REFERENCE OF CREATOR
-		this.creator.setLastReference(this.signature, db);
-	}
-
-	@Override
-	public void orphan(DBSet db) {
-
-		// NAME STORAGE UPDATE ORPHAN
-		// if (service == 10) {
-		// StorageUtils.processOrphan(getData(), signature, db);
-		// // BLOGPOST?
-		// } else {
-		// removeFromBlogMapOnDemand(db);
-		// }
-
-		// UPDATE CREATOR
-		this.creator.setConfirmedBalance(this.creator.getConfirmedBalance(db)
-				.add(this.fee), db);
-
-		// UPDATE REFERENCE OF CREATOR
-		this.creator.setLastReference(this.reference, db);
-	}
-
-	@Override
-	public PublicKeyAccount getCreator() {
-		return this.creator;
-	}
-
-	@Override
-	public List<Account> getInvolvedAccounts() {
-		List<Account> accounts = new ArrayList<Account>();
-
-		accounts.add(this.creator);
-
-		return accounts;
-	}
-
-	@Override
-	public boolean isInvolved(Account account) {
-		String address = account.getAddress();
-
-		if (address.equals(this.creator.getAddress())) {
-			return true;
-		}
-
-		return false;
-	}
-
-	@Override
-	public BigDecimal getAmount(Account account) {
-		String address = account.getAddress();
-
-		if (address.equals(this.creator.getAddress())) {
-			return BigDecimal.ZERO.setScale(8).subtract(this.fee);
-		}
-
-		return BigDecimal.ZERO.setScale(8);
-	}
-
 	public static byte[] generateSignature(DBSet db, PrivateKeyAccount creator,
 			int service, byte[] arbitraryData, BigDecimal fee, long timestamp) {
 		byte[] data = new byte[0];
@@ -387,226 +257,5 @@ public class ArbitraryTransactionV1 extends ArbitraryTransaction {
 		data = Bytes.concat(data, feeBytes);
 
 		return Crypto.getInstance().sign(creator, data);
-	}
-
-	@Override
-	public void addToCommentMapOnDemand(DBSet db) {
-
-		if (getService() == BlogUtils.COMMENT_SERVICE_ID) {
-			byte[] data = getData();
-			String string = new String(data, Charsets.UTF_8);
-
-			JSONObject jsonObject = (JSONObject) JSONValue.parse(string);
-			if (jsonObject != null) {
-				
-				String signatureOfCommentOpt = (String) jsonObject
-						.get(BlogPostResource.DELETE_KEY);
-				
-				//CHECK IF THIS IS A DELETE OR CREATE OF A COMMENT
-				if(StringUtils.isNotBlank(signatureOfCommentOpt))
-				{
-					BlogEntry commentEntryOpt = BlogUtils.getCommentBlogEntryOpt(signatureOfCommentOpt);
-					
-
-					if (commentEntryOpt != null) {
-						String creatorOfDeleteTX = getCreator().getAddress();
-						String creatorOfEntryToDelete = commentEntryOpt
-								.getCreator();
-
-							// OWNER IS DELETING OWN POST?
-							if (creatorOfDeleteTX
-									.equals(creatorOfEntryToDelete)) {
-								deleteCommentInternal(db, commentEntryOpt);
-								// BLOGOWNER IS DELETING POST
-							} else if (commentEntryOpt.getBlognameOpt() != null) {
-								Name name = db.getNameMap().get(
-										commentEntryOpt.getBlognameOpt());
-								if (name != null
-										&& name.getOwner().getAddress()
-												.equals(creatorOfDeleteTX)) {
-									deleteCommentInternal(db, commentEntryOpt);
-									
-								}
-							}
-
-					}
-				}else
-				{
-					String post = (String) jsonObject
-							.get(BlogPostResource.POST_KEY);
-					
-					String postid = (String) jsonObject
-							.get(BlogPostResource.COMMENT_POSTID_KEY);
-					
-					// DOES POST MET MINIMUM CRITERIUM?
-					if (StringUtils.isNotBlank(post)
-							&& StringUtils.isNotBlank(postid)) {
-						
-						db.getPostCommentMap().add(Base58.decode(postid),
-								getSignature());
-						db.getCommentPostMap().add(getSignature(), Base58.decode(postid));
-					}
-				}
-				
-				
-			}
-
-		}
-
-	}
-
-	private void addToBlogMapOnDemand(DBSet db) {
-
-		if (getService() == 777) {
-			byte[] data = getData();
-			String string = new String(data, Charsets.UTF_8);
-
-			JSONObject jsonObject = (JSONObject) JSONValue.parse(string);
-			if (jsonObject != null) {
-				String post = (String) jsonObject
-						.get(BlogPostResource.POST_KEY);
-
-				String blognameOpt = (String) jsonObject
-						.get(BlogPostResource.BLOGNAME_KEY);
-
-				String share = (String) jsonObject
-						.get(BlogPostResource.SHARE_KEY);
-
-				String delete = (String) jsonObject
-						.get(BlogPostResource.DELETE_KEY);
-
-				String author = (String) jsonObject
-						.get(BlogPostResource.AUTHOR);
-
-				boolean isShare = false;
-				if (StringUtils.isNotEmpty(share)) {
-					isShare = true;
-					byte[] sharedSignature = Base58.decode(share);
-					if (sharedSignature != null) {
-						db.getSharedPostsMap().add(sharedSignature, author);
-					}
-				}
-
-				if (StringUtils.isNotEmpty(delete)) {
-					BlogEntry blogEntryOpt = BlogUtils.getBlogEntryOpt(delete);
-
-					if (blogEntryOpt != null) {
-						String creatorOfDeleteTX = getCreator().getAddress();
-						String creatorOfEntryToDelete = blogEntryOpt
-								.getCreator();
-						if (blogEntryOpt != null) {
-
-							// OWNER IS DELETING OWN POST?
-							if (creatorOfDeleteTX
-									.equals(creatorOfEntryToDelete)) {
-								deleteInternal(db, isShare, blogEntryOpt);
-								// BLOGOWNER IS DELETING POST
-							} else if (author != null
-									&& blogEntryOpt.getBlognameOpt() != null) {
-								Name name = db.getNameMap().get(
-										blogEntryOpt.getBlognameOpt());
-								if (name != null
-										&& name.getOwner().getAddress()
-												.equals(creatorOfDeleteTX)) {
-									deleteInternal(db, isShare, blogEntryOpt);
-								}
-							}
-
-						}
-					}
-
-				} else {
-
-					// DOES POST MET MINIMUM CRITERIUM?
-					if (StringUtils.isNotBlank(post)) {
-
-						// Shares won't be hashtagged!
-						if (!isShare) {
-							List<String> hashTags = BlogUtils.getHashTags(post);
-							for (String hashTag : hashTags) {
-								db.getHashtagPostMap().add(hashTag,
-										getSignature());
-							}
-						}
-
-						db.getBlogPostMap().add(blognameOpt, getSignature());
-					}
-				}
-
-			}
-		}
-	}
-
-	@Override
-	public void deleteInternal(DBSet db, boolean isShare, BlogEntry blogEntryOpt) {
-		if (isShare) {
-			byte[] sharesignature = Base58.decode(blogEntryOpt
-					.getShareSignatureOpt());
-			db.getBlogPostMap().remove(blogEntryOpt.getBlognameOpt(),
-					sharesignature);
-			db.getSharedPostsMap().remove(sharesignature,
-					blogEntryOpt.getNameOpt());
-		} else {
-			// removing from hashtagmap
-			List<String> hashTags = BlogUtils.getHashTags(blogEntryOpt
-					.getDescription());
-			for (String hashTag : hashTags) {
-				db.getHashtagPostMap().remove(hashTag,
-						Base58.decode(blogEntryOpt.getSignature()));
-			}
-			db.getBlogPostMap().remove(blogEntryOpt.getBlognameOpt(),
-					Base58.decode(blogEntryOpt.getSignature()));
-		}
-	}
-	
-	@Override
-	public void deleteCommentInternal(DBSet db, BlogEntry commentEntry) {
-		
-		byte[] signatureOfComment = Base58.decode(commentEntry.getSignature());
-			byte[] signatureOfBlogPostOpt = db.getCommentPostMap().get(Base58.decode(commentEntry.getSignature()));
-			// removing from hashtagmap
-			
-			if(signatureOfBlogPostOpt != null)
-			{
-				db.getPostCommentMap().remove(signatureOfBlogPostOpt, signatureOfComment);
-				db.getCommentPostMap().remove(signatureOfComment);
-				
-			}
-	}
-
-	// TODO implement readd delete if orphaned!
-	@SuppressWarnings("unused")
-	private void removeFromBlogMapOnDemand(DBSet db) {
-		if (getService() == 777) {
-			byte[] data = getData();
-			String string = new String(data, Charsets.UTF_8);
-
-			JSONObject jsonObject = (JSONObject) JSONValue.parse(string);
-			if (jsonObject != null) {
-				String blognameOpt = (String) jsonObject
-						.get(BlogPostResource.BLOGNAME_KEY);
-
-				String share = (String) jsonObject
-						.get(BlogPostResource.SHARE_KEY);
-
-				String author = (String) jsonObject
-						.get(BlogPostResource.AUTHOR);
-
-				if (StringUtils.isNotEmpty(share)) {
-					byte[] sharedSignature = Base58.decode(share);
-					if (sharedSignature != null) {
-						db.getSharedPostsMap().remove(sharedSignature, author);
-					}
-				}
-
-				db.getBlogPostMap().remove(blognameOpt, getSignature());
-
-			}
-		}
-	}
-
-	@Override
-	public List<Payment> getPayments() {
-		return new ArrayList<Payment>();
 	}
 }
